pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "A nano Claude Code-like agent using AWS Bedrock Converse API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["agents"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
